pyproject.toml
[project]
name = "mysite"
version = "0.1.0"
description = ""
authors = [
    {name = "kvi",email = "kvi92@ya.ru"}
]

# По умолчанию предлагает ограничить нижнюю границу. Работать условие будет до "4.0":
# requires-python = "^3.13"
# Ограничить только в пределах патч-версии можно несколькими вариантами:
# requires-python = "~3.13"
# requires-python = "3.13.*"
# requires-python = ">=3.13, <3.14"
# Так как целевое приложение соберается в Docker (версия python определится образом), решено оставить "по умолчанию":
requires-python = "^3.13"

# Для полного воспроизведения настроек зафиксированы. Оставлен рекомендованный диапазон версий.
# Это позволит устанавливать версии бибилотек, в которых гарантирована совместимость.
dependencies = [
    "django (>=6.0.3,<7.0.0)",
    "django-rest-framework (>=0.1.0,<0.2.0)",
    "pillow (>=12.1.1,<13.0.0)",
    "django-debug-toolbar (>=6.2.0,<7.0.0)",
    "django-filter (>=25.2,<26.0)",
    "drf-spectacular (>=0.29.0,<0.30.0)",
    "docutils (>=0.22.4,<0.23.0)",
    "gunicorn (>=25.1.0,<26.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
